리팩터: DBErrorWrapper를 유틸로 옮김
